+++ src/elements/length-determinant.ts
@@ -1,7 +1,8 @@
+import { BitstreamElement } from "./element";
 import { FieldDefinition } from "./field-definition";
 
 /**
  * Determines the bitlength of a certain field. Can be a number or a function that dynamically
  * determines the value based on the current context.
  */
- export type LengthDeterminant = number | ((instance : any, f : FieldDefinition) => number);
+ export type LengthDeterminant<T extends BitstreamElement> = number | ((instance : T, f : FieldDefinition) => number);