+++ src/elements/reserved.ts
@@ -4,6 +4,7 @@ import { FieldOptions } from "./field-options";
 import { LengthDeterminant } from "./length-determinant";
 import { Field } from "./field";
 import { BitstreamElement } from "./element";
+import { InferredPropertyDecorator, PropType } from "./decorators";
 
 /**
  * Used to mark a specific field as reserved. The value in this field will be read, but will not be 
@@ -12,22 +13,22 @@ import { BitstreamElement } from "./element";
  * @param length The bitlength determinant
  * @param options Options related to this reserved field
  */
- export function Reserved<T extends BitstreamElement>(length : LengthDeterminant, options? : FieldOptions<T>) {
-    if (!options)
-        options = {};
 
-    options.isIgnored = true;
-    options.writtenValue = (instance, field : FieldDefinition) => {
-        if (field.type === Number) {
-            let currentLength = resolveLength(field.length, instance, field);
-            return Math.pow(2, currentLength) - 1;
-        }
-    };
-
-    let decorator = Field(length, options);
+export function Reserved<T extends BitstreamElement>(length : LengthDeterminant<T>, options : Omit<FieldOptions<T, number>, 'writtenValue' | 'isIgnored'> = {}): InferredPropertyDecorator<T>
+ {
     return (target : T, fieldName : string | symbol) => {
         fieldName = Symbol(`[reserved: ${typeof length === 'number' ? `${length} bits` : `dynamic`}]`);
         Reflect.defineMetadata('design:type', Number, target, fieldName);
-        return decorator(target, fieldName);
+
+        return Field<T, number>(length, {
+            ...options,
+            isIgnored: true,
+            writtenValue: (instance, field : FieldDefinition) => {
+                if (field.type === Number) {
+                    let currentLength = resolveLength(field.length, instance, field);
+                    return Math.pow(2, currentLength) - 1;
+                }
+            }
+        })(target, fieldName);
     }
 }